[Migration 4-6] Move rollup/shibarium to client/features/rollup/shibarium/
Closes #3431
