+++ client/features/rollup/shibarium/pages/deposits/DepositsListItem.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import type { ShibariumDepositsItem } from 'types/api/shibarium';
+import type { ShibariumDepositsItem } from 'client/features/rollup/shibarium/types/api';
 
 import AddressStringOrParam from 'client/slices/address/components/entity/AddressStringOrParam';
 import TxEntity from 'client/slices/tx/components/entity/TxEntity';
