+++ client/features/rollup/shibarium/pages/deposits/__screenshots__/ShibariumDeposits.pw.tsx_default_base-view-mobile-1.png
