+++ client/features/rollup/shibarium/pages/deposits/__screenshots__/ShibariumDeposits.pw.tsx_mobile_base-view-mobile-1.png
