+++ client/features/rollup/shibarium/pages/withdrawals/WithdrawalsTable.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import type { ShibariumWithdrawalsItem } from 'types/api/shibarium';
+import type { ShibariumWithdrawalsItem } from 'client/features/rollup/shibarium/types/api';
 
 import { AddressHighlightProvider } from 'client/slices/address/contexts/address-highlight';
