+++ client/features/rollup/shibarium/pages/withdrawals/__screenshots__/ShibariumWithdrawals.pw.tsx_mobile_base-view-mobile-1.png
