+++ client/features/rollup/shibarium/stubs.ts
@@ -1,4 +1,4 @@
-import type { ShibariumDepositsItem, ShibariumWithdrawalsItem } from 'types/api/shibarium';
+import type { ShibariumDepositsItem, ShibariumWithdrawalsItem } from 'client/features/rollup/shibarium/types/api';
 
 import { ADDRESS_PARAMS } from 'client/slices/address/stubs/address-params';
 import { TX_HASH } from 'client/slices/tx/stubs/tx';
