--- ui/withdrawals/shibarium/WithdrawalsListItem.tsx
@@ -1,83 +1,0 @@
-// SPDX-License-Identifier: LicenseRef-Blockscout
-
-import React from 'react';
-
-import type { ShibariumWithdrawalsItem } from 'types/api/shibarium';
-
-import AddressStringOrParam from 'client/slices/address/components/entity/AddressStringOrParam';
-import BlockEntity from 'client/slices/block/components/entity/BlockEntity';
-import TxEntity from 'client/slices/tx/components/entity/TxEntity';
-
-import TxEntityL1 from 'client/features/rollup/common/components/TxEntityL1';
-
-import config from 'configs/app';
-import { layerLabels } from 'lib/rollups/utils';
-import ListItemMobileGrid from 'ui/shared/ListItemMobile/ListItemMobileGrid';
-import TimeWithTooltip from 'ui/shared/time/TimeWithTooltip';
-
-const feature = config.features.rollup;
-
-type Props = { item: ShibariumWithdrawalsItem; isLoading?: boolean };
-
-const WithdrawalsListItem = ({ item, isLoading }: Props) => {
-  if (!(feature.isEnabled && feature.type === 'shibarium')) {
-    return null;
-  }
-
-  return (
-    <ListItemMobileGrid.Container>
-
-      <ListItemMobileGrid.Label isLoading={ isLoading }>{ layerLabels.current } block No</ListItemMobileGrid.Label>
-      <ListItemMobileGrid.Value>
-        <BlockEntity
-          number={ item.l2_block_number }
-          isLoading={ isLoading }
-          textStyle="sm"
-          fontWeight={ 600 }
-        />
-      </ListItemMobileGrid.Value>
-
-      <ListItemMobileGrid.Label isLoading={ isLoading }>{ layerLabels.current } txn hash</ListItemMobileGrid.Label>
-      <ListItemMobileGrid.Value>
-        <TxEntity
-          isLoading={ isLoading }
-          hash={ item.l2_transaction_hash }
-          textStyle="sm"
-          truncation="constant_long"
-        />
-      </ListItemMobileGrid.Value>
-
-      <ListItemMobileGrid.Label isLoading={ isLoading }>{ layerLabels.parent } txn hash</ListItemMobileGrid.Label>
-      <ListItemMobileGrid.Value>
-        <TxEntityL1
-          isLoading={ isLoading }
-          hash={ item.l1_transaction_hash }
-          textStyle="sm"
-          truncation="constant_long"
-          noCopy
-        />
-      </ListItemMobileGrid.Value>
-
-      <ListItemMobileGrid.Label isLoading={ isLoading }>User</ListItemMobileGrid.Label>
-      <ListItemMobileGrid.Value>
-        <AddressStringOrParam
-          address={ item.user }
-          isLoading={ isLoading }
-          noCopy
-          truncation="constant"
-        />
-      </ListItemMobileGrid.Value>
-      <ListItemMobileGrid.Label isLoading={ isLoading }>Age</ListItemMobileGrid.Label>
-      <ListItemMobileGrid.Value>
-        <TimeWithTooltip
-          timestamp={ item.timestamp }
-          isLoading={ isLoading }
-          display="inline-block"
-        />
-      </ListItemMobileGrid.Value>
-
-    </ListItemMobileGrid.Container>
-  );
-};
-
-export default WithdrawalsListItem;
